smart-pantry_py/requirements.txt
flask
scipy
requests
awsiotsdk
pandas